requirements.txt
streamlit>=1.44.1
pandas
pydeck
requests>=2.32.3
python-dotenv>=1.1.0
plotly
openai>=1.75.0
